[refactor] 마이페이지 코드리뷰 피드백
