+++ src/pages/MyPage/components/Mailbox/index.tsx
@@ -1,11 +1,33 @@
-import Letters from '@/pages/MyPage/components/Mailbox/Letters'
+import Letter from '@/pages/MyPage/components/Mailbox/Letter'
 
-interface MailProps {
+interface LettersProps {
+  className?: string
+  id?: string
   status: string
 }
 
-const Mailbox = ({ status }: MailProps) => {
-  return <Letters status={status} />
+const Letters = ({ status }: LettersProps) => {
+  return (
+    <div className="px-4 py-5">
+      {status === 'received' ? (
+        <div className="mx-2.5 flex flex-col gap-4 overflow-y-scroll">
+          <Letter />
+          <Letter />
+          <Letter />
+          <Letter />
+          <Letter />
+        </div>
+      ) : (
+        <div className="mx-2.5 flex flex-col gap-4">
+          <Letter />
+          <Letter />
+          <Letter />
+          <Letter />
+          <Letter />
+        </div>
+      )}
+    </div>
+  )
 }
 
-export default Mailbox
+export default Letters
